refactor(ActivateUser): update the directory structure
